Switch to SSR in HomePage
