+++ components/Home/Newsletter.tsx
@@ -1,21 +1,39 @@
-import { FormEvent, useState } from "react";
+"use client";
+
+import { connect } from "react-redux";
+import { FormEvent, useEffect, useState } from "react";
+import { usePathname } from "next/navigation";
+import { useCookies } from "react-cookie";
 import Image from "next/image";
-import { SubscribeMessage } from "../../app/home-page";
+import Modal from "./Modal";
+import { subscribeUser } from "lib/subscribeUser";
+import NotificationBanner from "components/Shared/NotificationBanner";
+
+type SubscribeMessage = { isSuccess: boolean; message: string };
 
 interface Props {
-  handleSubscribe: (
-    e: FormEvent<HTMLFormElement>,
-    formData: NewsletterFormData
-  ) => Promise<SubscribeMessage>;
+  notification: { name: string; message: string; status: string };
 }
 
-const Newsletter = ({ handleSubscribe }: Props) => {
+const Newsletter = ({ notification }: Props) => {
+  const [modalVisible, setModalVisible] = useState<boolean>(false);
   const [formData, setFormData] = useState<NewsletterFormData>({
     name: "",
     email: "",
   });
   const [isValid, setIsValid] = useState<boolean>(true);
   const [errorMessage, setErrorMessage] = useState("");
+
+  const pathName = usePathname();
+  const [cookies, setCookie] = useCookies(["access_token"]);
+
+  useEffect(() => {
+    const access_token = pathName?.match(/\#(?:access_token)\=([\S\s]*?)\&/);
+
+    if (access_token && access_token.length > 1) {
+      setCookie("access_token", access_token[1]);
+    }
+  }, [pathName, setCookie]);
 
   const onSubmit = async (e) => {
     const { isSuccess, message } = await handleSubscribe(e, formData);
@@ -29,8 +47,36 @@ const Newsletter = ({ handleSubscribe }: Props) => {
     }
   };
 
+  async function handleSubscribe(
+    e: FormEvent<HTMLFormElement>,
+    { name, email }: NewsletterFormData
+  ): Promise<SubscribeMessage> {
+    e.preventDefault();
+    if (name && email) {
+      try {
+        await subscribeUser(name, email);
+
+        setModalVisible(true);
+        return {
+          isSuccess: true,
+          message: "",
+        };
+      } catch (error) {
+        return {
+          isSuccess: false,
+          message: "ERROR: Email already exists",
+        };
+      }
+    } else
+      return {
+        isSuccess: false,
+        message: "ERROR: Name or Email Cannot be Empty",
+      };
+  }
+
   return (
     <div className="mt-14 relative z-0">
+      <Modal modalVisible={modalVisible} setModalVisible={setModalVisible} />
       <Image
         src="/newsbg.png"
         alt="BG Newsletter Image"
@@ -91,8 +137,20 @@ const Newsletter = ({ handleSubscribe }: Props) => {
           )}
         </div>
       </div>
+      {notification && (
+        <NotificationBanner
+          message={notification.message}
+          status={notification.status}
+        />
+      )}
     </div>
   );
 };
 
-export default Newsletter;
+const mapStateToPros = (state) => {
+  return {
+    notification: state.main?.notification,
+  };
+};
+
+export default connect(mapStateToPros)(Newsletter);
