+++ components/Home/VerseOfDay.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useEffect, useState } from "react";
 import Link from "next/link";
 import { Skeleton } from "../Shared/Skeleton";
